requirements.txt
qiskit
qiskit-aer
pandas
matplotlib
numpy
python-dotenv